Move permission constants to new file, add displayName to policies
