+++ packages/shared/src/permissions/index.ts
@@ -1,2 +1,3 @@
 export * from "./permissions.utils";
 export * from "./permissionsClass";
+export * from "./permissions.constants";
